fix so server will be killed on close

+++ server.js
@@ -50,7 +50,7 @@ var runServer = {
     },
     // kill process for restart
     kill: function (signal) {
-        this.child.kill(signal)
+        if(this.child) this.child.kill(signal)
     }
 
 }
